refacto: Split model in two parts
The model file was getting a bit too crowded, and I'm planning on adding
a Shipyard subclass to the Waypoint, so it made sense to split it.

I've decided to split it in two for now: the things related mainly to
the agent (contracts, etc) and the thing related to navigation and
waypoints.