refacto: Use repositories in purchase_ship
This brings the purchase_ship function to the same level of abstraction
than other functions in the orchestration/ folder: they do not use our
entities directly as arguments, but instead require repositories as
arguments.

I'm not 100% of this decision, but let's at least be consistent. If it
turns out not to be the best choice, I'll refactor them later on.