feat(agent): Add a new implementation
This implementation for the agent repository will be used in the future
integration test.

It does not rely on the file at all: the token is not saved. The aim is
to generate a "throwaway" agent, to be used in the integration test.
This is why a random string is suffixed to the agent symbol: to allow
running the test several times in a row, without interfering with the
"real" agent who's token is saved in the file.