chore(test): Add a first integration test
This test runs the whole tutorial to make sure the various Http
repositories are properly implemented. To avoid messing with the
"normal" agent, we use the new PureHttpAgentRepository for this test.

This test will be extended as my implementation of the tutorial
progresses, to cover the whole tutorial in the end.