refactor(agent): Change in-memory agent init
This is done to try and decouple our test from our domain so we're freer
to refactor our domain later on (see Chapter 5 in Cosmic Python).

A first step is to create an agent factory method in our in-memory
repository to avoid having to create agents in the test code, and to
avoid having implicit agent creation code in our repository. Now, the
responsibilities are split cleanly.

This also means that if/when we'll change the implementation of our
Agent class, we will only have to change the Factory method instead of
going through all our tests and manually fixing them one by one.