refactor(contract): Change in-memory contract repo
This is done to try and decouple our test from our domain so we're freer
to refactor our domain later on (see Chapter 5 in Cosmic Python).

We add a new Factory method to the Contract repository module that
produces an InMemoryContractRepository. We also add a new method to the
InMemoryContractRepository: it allows us to create a contract using only
base data types.

This means that if/when we'll change the implementation of our
Contract class, we will only have to change the Factory method instead of
going through all our tests and manually fixing them one by one.