refactor: Various contract improvements
We add one missing orchestration function: the accept_contract() method.
For some reason, this was done "manually" in the code instead of calling
an orchestration method. This is now more in line with how the rest of
the code behaves.

We have also modified the tests to use this new method instead of
calling to our domain directly, allowing decoupling from those tests
with our domain.