refactor(waypoints): Add the System entitiy
I was unsure of how to refactor the waypoints repository to decouple my
tests from it. Upon thinking about it, I realised that my current
implementation was wonky because it was mixing different things in the
same entity. The previous waypoint entity (and the associated
repository) were used to interact both at system level and at waypoint
level.

This made it complex to cleanly set up tests and decouple them from the
domain because depending on the tests, I needed to provide either one
waypoint or a full system. By introducing this new entity, I hope that
I'll be able to decouple my tests from the domain.